First Implementation of AC
I've noticed some people have been using autoclickers as a way to cheese through the game. This is the first implementation to handle those people. It's not perfect. Infact I have another solution in the works. I'll be working on that as I prefer that one instead of this one.